#45 feat: 게임방 생성 시 RedisTemplate으로 게임 중인 멤버 현황 관리 migration

+++ src/main/java/jungle/HandTris/application/impl/GameRoomServiceImpl.java
@@ -8,39 +8,50 @@ import jungle.HandTris.domain.GameStatus;
 import jungle.HandTris.domain.exception.GameRoomNotFoundException;
 import jungle.HandTris.domain.exception.ParticipantLimitedException;
 import jungle.HandTris.domain.exception.PlayingGameException;
-import jungle.HandTris.domain.repo.GameRoomRepository;
 import jungle.HandTris.presentation.dto.request.GameRoomDetailReq;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.UUID;
-
+import java.util.stream.Collectors;
 
 @Service
 @Transactional
 @RequiredArgsConstructor
 public class GameRoomServiceImpl implements GameRoomService {
-    private final GameRoomRepository gameRoomRepository;
-    // 캐시 사용을 위한 주입
+
+    private final RedisTemplate<String, GameRoom> redisTemplate;
     private final GameMemberService gameMemberService;
+    private static final String GAME_ROOM_KEY_PREFIX = "gameMember:";
 
     public List<GameRoom> getGameRoomList() {
-        return gameRoomRepository.findAllByGameStatusNotPlaying();
+        return redisTemplate.keys(GAME_ROOM_KEY_PREFIX + "*").stream()
+                .map(key -> redisTemplate.opsForValue().get(key))
+                .filter(gameRoom -> gameRoom != null && gameRoom.getGameStatus() != GameStatus.PLAYING)
+                .collect(Collectors.toList());
     }
 
     public UUID createGameRoom(GameRoomDetailReq gameRoomDetailReq) {
         GameRoom createdGameRoom = new GameRoom(gameRoomDetailReq);
-        gameRoomRepository.save(createdGameRoom);
+        String gameRoomKey = GAME_ROOM_KEY_PREFIX + createdGameRoom.getUuid().toString();
+
+        redisTemplate.opsForValue().set(gameRoomKey, createdGameRoom);
 
         // Redis에 방Id와 유저Id 매핑 후 저장
-        gameMemberService.addMemberToRoom(createdGameRoom.getUuid().toString(), "JWT에서 얻은 유저ID");
+//        gameMemberService.addMemberToRoom(createdGameRoom.getUuid().toString(), "JWT에서 얻은 유저ID");
 
         return createdGameRoom.getUuid();
     }
 
     public GameRoom enterGameRoom(long gameId) {
-        GameRoom gameRoom = gameRoomRepository.findById(gameId).orElseThrow(GameRoomNotFoundException::new);
+        String gameRoomKey = GAME_ROOM_KEY_PREFIX + gameId;
+        GameRoom gameRoom = redisTemplate.opsForValue().get(gameRoomKey);
+
+        if (gameRoom == null) {
+            throw new GameRoomNotFoundException();
+        }
 
         if (gameRoom.getGameStatus() == GameStatus.PLAYING) {
             throw new PlayingGameException();
@@ -51,7 +62,7 @@ public class GameRoomServiceImpl implements GameRoomService {
         }
 
         gameRoom.enter();
-        gameRoomRepository.save(gameRoom); // TODO 강호형쪽에서 삭제예정
+        redisTemplate.opsForValue().set(gameRoomKey, gameRoom);
 
         // Redis에 방Id와 유저Id 매핑 후 저장
         gameMemberService.addMemberToRoom(gameRoom.getUuid().toString(), "JWT에서 얻은 유저ID");
@@ -60,7 +71,12 @@ public class GameRoomServiceImpl implements GameRoomService {
     }
 
     public GameRoom exitGameRoom(long gameId) {
-        GameRoom gameRoom = gameRoomRepository.findById(gameId).orElseThrow(GameRoomNotFoundException::new);
+        String gameRoomKey = GAME_ROOM_KEY_PREFIX + gameId;
+        GameRoom gameRoom = redisTemplate.opsForValue().get(gameRoomKey);
+
+        if (gameRoom == null) {
+            throw new GameRoomNotFoundException();
+        }
 
         if (gameRoom.getGameStatus() == GameStatus.PLAYING) {
             throw new PlayingGameException();
@@ -74,13 +90,18 @@ public class GameRoomServiceImpl implements GameRoomService {
             deleteGameRoom(gameId);
             return gameRoom;
         }
-        gameRoomRepository.save(gameRoom); // TODO 강호형쪽에서 삭제예정
+        redisTemplate.opsForValue().set(gameRoomKey, gameRoom);
         return gameRoom;
     }
 
-    public void deleteGameRoom(long gameId) { // TODO 강호형쪽에서 삭제예정
-        GameRoom gameRoom = gameRoomRepository.findById(gameId).orElseThrow(GameRoomNotFoundException::new);
-        gameRoomRepository.delete(gameRoom);
+    public void deleteGameRoom(long gameId) {
+        String gameRoomKey = GAME_ROOM_KEY_PREFIX + gameId;
+        GameRoom gameRoom = redisTemplate.opsForValue().get(gameRoomKey);
+
+        if (gameRoom == null) {
+            throw new GameRoomNotFoundException();
+        }
+        redisTemplate.delete(gameRoomKey);
 
         // Redis에서 방 삭제
         gameMemberService.deleteRoom(gameRoom.getUuid().toString());
